json: Test and code cleanup
