+++ modules/json/src/helpers/instantiate-class.js
