+++ test/apps/json-pure-js/app.js
@@ -87,7 +87,7 @@ function initializeApp() {
 const jsonConverter = new JSONConverter({configuration: JSON_CONFIGURATION});
 
 function setJSON(json, updateEditor = true) {
-  const deckProps = jsonConverter.convertJson(json);
+  const deckProps = jsonConverter.convert(json);
   deck.setProps(deckProps);
   setMapProps(deckProps);
   map._container.style.visibility = deckProps.map ? 'visible' : 'hidden';
