+++ test/modules/json/data/deck-props.json
@@ -1,0 +1,37 @@
+{
+  "initialViewState": {
+    "longitude": -122.45,
+    "latitude": 37.8,
+    "zoom": 12
+  },
+  "mapStyle": {},
+  "views": [
+    {
+      "type": "MapView",
+      "height": "50%",
+      "controller": true
+    },
+    {
+      "type": "FirstPersonView",
+      "y": "50%",
+      "height": "50%"
+    }
+  ],
+  "layers": [
+    {
+      "type": "ScatterplotLayer",
+      "data": [{"position": [-122.45, 37.8]}],
+      "getPosition": "position",
+      "getFillColor": [255, 0, 0, 255],
+      "getRadius": 1000
+    },
+    {
+      "type": "TextLayer",
+      "data": [[-122.45, 37.8]],
+      "getPosition": "-",
+      "elevationRange": [0, 3000],
+      "getText": "Hello World",
+      "billboard": false
+    }
+  ]
+}
