+++ test/modules/json/index.js
@@ -1,11 +1,11 @@
 import './utils/get.spec';
 import './utils/shallow-equal-objects.spec';
 
-import './lib/helpers/deck-prop-types.spec';
-import './lib/helpers/parse-expression-string.spec.js';
-import './lib/helpers/convert-functions.spec';
+import './helpers/deck-prop-types.spec';
+import './helpers/parse-expression-string.spec';
+import './helpers/convert-functions.spec';
 
-import './lib/json-converter.spec.js';
+import './json-configuration.spec';
+import './json-converter.spec';
 
-// import './no-longer-supported/deck-json-converter.spec';
-// import './no-longer-supported/json-render.spec';
+import './json-render.spec';
